+++ tests/test_account.py
@@ -152,6 +152,23 @@ class TestAccount(unittest.TestCase):
         with self.assertRaises(CanvasException):
             self.account.delete()
 
+    # delete_admin()
+    def test_delete_admin(self, m):
+        register_uris({"account": ["delete_admin"]}, m)
+
+        user_id = 123
+        deleted_admin = self.account.delete_admin(user=user_id)
+
+        self.assertIsInstance(deleted_admin, Admin)
+        self.assertTrue(hasattr(deleted_admin, "id"))
+        self.assertTrue(hasattr(deleted_admin, "role"))
+        self.assertTrue(hasattr(deleted_admin, "role_id"))
+        self.assertTrue(hasattr(deleted_admin, "workflow_state"))
+        self.assertEqual(deleted_admin.user["login_id"], "jdoe")
+        self.assertEqual(deleted_admin.role, "AccountAdmin")
+        self.assertEqual(deleted_admin.role_id, 1)
+        self.assertEqual(deleted_admin.workflow_state, "deleted")
+
     # delete_user()
     def test_delete_user_id(self, m):
         register_uris({"account": ["delete_user"]}, m)